# aws/signer/v4: Remove string concats from stripExcessSpaces