Removes the string concats from stripExcessSpaces that was causing many
allocs for each signed request.

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkStripExcessSpaces-4     18815         542           -97.12%

benchmark                        old allocs     new allocs     delta
BenchmarkStripExcessSpaces-4     406            5              -98.77%

benchmark                        old bytes     new bytes     delta
BenchmarkStripExcessSpaces-4     17616         144           -99.18%
```

Fix #712